model-service/requirements.txt
transformers
grpcio-tools